More foundational framework

+++ NanoCart/Entities/Company.cs
@@ -2,6 +2,6 @@ namespace NanoCart.Entities;
 
 public class Company
 {
-    private int id { get; set; }
+    private long id { get; set; }
     private string name { get; set; }
 }
